requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

# LangChain + integrations
langchain==0.2.12
langchain-community==0.2.10
langchain-core==0.2.34
langchain-text-splitters==0.2.2

# Vector DB + embeddings
#chromadb==0.5.5
chromadb==0.5.5
sentence-transformers==3.0.1

# Loaders & utils
pypdf==4.3.1
asteval==0.9.31

streamlit==1.37.1
